==> 2_create_analysisdata.rmd <==
---
title: "Karlbom: Create analysisdata"
author: "John Moshtaghi-Svensson"
output: 
    html_document:
        highlight: pygments
        theme: sandstone
        css: style.css
        toc: true
        code_folding: show
editor_options: 
  chunk_output_type: inline
---

__Datum__: `r lubridate::today()`

```{r setup, include=FALSE, message=FALSE, warning=FALSE, echo=FALSE}
library(data.table)
library(dplyr)
library(stringr)
library(glue)
library(janitor)
library(here)
library(scales)
library(ggplot2)
library(lubridate)
library(rlang)
library(plotly)
library(tidyr)

source(".Rprofile")
source("crcbase_utils/helpers.R")
source("crcbase_utils/crud/setup_project_crud.R")
source("crcbase_utils/process/fix_scrcr_dubletter.R")
source("crcbase_utils/process/process_scrcr.R")
source("crcbase_utils/process/process_lisa.R")
# Loads a list called definitions containtin procedure_codes
source("vars/definitions.R")

test_for_k()

theme_set(theme_bw())

knitr::opts_chunk$set(
    # root.dir = '~/R/project',
    warning = FALSE,
    message = FALSE,
    warning = FALSE,
    echo = TRUE,
    include = TRUE
)

if (!exists("params")) {
    params <- list(
        end_of_study = "2016-12-31",
        data_date = "20220207", # when was the original data created
        project_name = project_name # from .Renviron, loaded in setup_project_crud.R
    )
}
```

# Description
Combines all data tables in project/data/original into one dataset and create all necessary new variables

# 1. Datasets
Studypopulation and migrations were created in `programs/crcbase-karlbom/1_create_original_data.R`
All of the sas files created in `programs/uk_20210609.sas` uses studypopulation to filter on lopnr

```{r read all data}
# Charlson for population
cci_exposed_raw <- read_from_project_original(glue("uk_cci_{params$data_date}.csv"))
cci_comparators_raw <- read_from_project_original(glue("uk_cci_comparators_{params$data_date}.csv"))

# Deaths for population
cdr_raw <- read_from_project_original(glue("uk_cdr_{params$data_date}.csv"))

# all iipr rows (before and after diagdate) indicating ileus, pain or hernia
# - ileus ("K565" "K566" "K567")
# - ab_pain (R10.0, R10.3, R10.4)
# - ab_hernia (K40.0, K40.3, K41.0, K41.3, K42.0, K43.0, K43.3, K43.6, K44.0, K45.0, K46.0)
ileus_raw <- read_from_project_original(glue("uk_ileus_{params$data_date}.csv"))

# all migrations for poulation
migrations_raw <- read_from_project_original(glue("uk_migrations_{params$data_date}.csv"))

# scrcr for population
scrcr_first <- read_from_project_original(glue("uk_scrcr_first_{params$data_date}.csv"))
scrcr_clean <- read_from_project_original(glue("uk_scrcr_clean_{params$data_date}.csv"))
scrcr_processed <- process_scrcr(scrcr_clean)

# all ipr row with ip codes specified in study plan (lots of codes, )
# Used to identify:
#- stoma closure IPR ONLY (op=JFG00, JFG10, JFG20, JFG23, JFG29, JFG26,JFG30, JFG33, JFG36)
surgeries_raw <- read_from_project_original(glue("uk_surgeries_{params$data_date}.csv"))

# Previous colorectal cancer (c18-c20) used for exclusion
previous_cr_cancer_raw <- read_from_project_original(glue("uk_previous_crc_{params$data_date}.csv"))
previous_cr_cancer <- previous_cr_cancer_raw[diagdate_scr < diagdate_scrcr]

# previous abdominal surgeries
previous_abd_surgeries_raw <- read_from_project_original(glue("uk_abd_surgeries_before_diagdate_{params$data_date}.csv"))

# first occurrence of ("C784" "C785" "C786" "C787") in patreg after diagdate_scrcr
# used for cencoring, i think
relapse_raw <- read_from_project_original(glue("uk_relapse_pr_{params$data_date}.csv"))

# Abdominal hernia and abdominal pain from IPR
abd_hernia_raw <- read_from_project_original(glue("uk_abd_hernia_{params$data_date}.csv"))

# Lisa for educational level
lisa_raw <- read_from_project_original(glue("uk_lisa_{params$data_date}.csv"))

# Study population
studypop_raw <- read_from_project_original("studypopulation.csv") # Created from studypopulation_comparators.sas7bdat and scrcr after inclusion and exlusion criteria
```

```{r import ables description}
import_tables <- list(
    cci_exposed_raw = cci_exposed_raw,
    cci_comparators_raw = cci_comparators_raw,
    cdr_raw = cdr_raw,
    ileus_raw = ileus_raw,
    migrations_raw = migrations_raw,
    scrcr_first = scrcr_first,
    scrcr_clean = scrcr_clean,
    surgeries_raw = surgeries_raw,
    previous_cr_cancer = previous_cr_cancer,
    previous_abd_surgeries_raw = previous_abd_surgeries_raw,
    relapse_raw = relapse_raw,
    abd_hernia_raw = abd_hernia_raw,
    studypop_raw = studypop_raw
)

studypop <- studypop_raw
exp_lopnr <- studypop[case == "exposed"]$lopnr
comp_lopnr <- studypop[case == "comparators"]$lopnr

purrr::imap_dfr(import_tables, function(tbl, name) {
    if ("data.fram" %in% class(tbl)) {
        message(glue("{name} is not a table. Returning NULL"))
        return(NULL)
    }
    tibble::tibble(
        name = name,
        ncol = ncol(tbl),
        nrow = nrow(tbl),
        n_unique_lopnr = length(unique(tbl$lopnr)),
        pop_in = scales::percent(mean(unique(studypop$lopnr) %in% tbl$lopnr)),
        exp_in = scales::percent(mean(unique(exp_lopnr) %in% tbl$lopnr)),
        comp_in = scales::percent(mean(unique(comp_lopnr) %in% tbl$lopnr)),
        has_index_date = "index_date" %in% colnames(tbl),
        has_match_date = "match_date" %in% colnames(tbl),
        has_diagdate = "diagdate_scrcr" %in% colnames(tbl),
    )
})
```

# 2. STUDY POPULATION

Most inclusion/exclusion criterion are alrady applied in `1_create_original_data.rmd`.

```{r studypopulation size}
studypop[,.N, by = "case"]
```

Applies in this file:

- previous colorectal cancer

Because previous_cr_cancer can only be defined after the studypopulation is.

### Exclusion criteria

#### 1. Previous Colorectal cancer

Identified from the swedish cancer register. Any before `diagdate_scrcr` plus 3months _a bit arbitrary but almost all have a cancer diagnosis before diagdate_scrcr. Im guessing it is  the same one. Need a washout period_
codes C18, C19, C20

- This is applied to the whole population (exposed and comparators)

```{r previous colorectal cancer}

previous_cr_cancer[, diagdate_scr_3m := diagdate_scr %m+% months(3)]
previous_cr_cancer[, timediff_m := lubridate::time_length(as_date(diagdate_scrcr) - as_date(diagdate_scr), unit = "months")]
prev_crc <- previous_cr_cancer[timediff_m > 3, .(lopnr, diagdate_scr, diagdate_scrcr, timediff_m)]

prev_crc_first <- prev_crc[, .(diagdate_crc = min(diagdate_scr), timediff_m = min(timediff_m)), by = c("lopnr", "diagdate_scrcr")]

#  plot distribution
qplot(prev_crc_first$timediff_m) + ggtitle("months since previous colorectal cancer")

# add to exposed statsu
prev_crc_first <- left_join(prev_crc_first, studypop, by = c("lopnr"))
# n exluded
prev_crc_first[, .N, by = "case"]

studypop <- studypop[!lopnr %in% prev_crc_first$lopnr]
```


### Exposed

Cancer surgery (bowel resection[procedure_type==3]) after colorectal cancer. Identified from SCRCR

```{r study population}
# studypop contains lopnr, match_date, strata, case, birthyear and some other stuff
# add all of the scrcr variables to exposed
pop <- left_merge(
    studypop,
    scrcr_first,
    by = c("lopnr", "diagdate_scrcr", "proceduredate")
) %>%
    select(lopnr, strata, case, match_date, proceduredate, birthyear, everything())
```

### Comparators

General population comparators. No information from SCRCR is available for comparators.

```{r}
pop[, .N, by = "case"]
# number of comparators per exposed
n_comp <- pop[, .(exposed = sum(case == "exposed"), comparators = sum(case == "comparators")), by = "strata"]
n_comp[, .N, by = "comparators"] %>% adorn_totals()
```

__Remove comparators without exposed__
```{r remove comparators without exposed}
strata_without_exposed <- n_comp[exposed == 0]$strata
length(strata_without_exposed)
pop <- pop[!strata %in% strata_without_exposed]
pop[, .N, by = "case"]
```


# 3. Variables

### Start of follow up

`index_date` = `procedure_date` from scrcr

```{r index date}
# Already added to studypopulation and pop data.frames
ggplot(pop, aes(x = index_date)) +
    geom_histogram(binwidth = 30) +
    facet_grid(~case)
```



### Cencoring factors

1. #### Death date

Added from CDR. Used for cencoring.

```{r add CDR}
# cdr_raw contains one row per index_date....only need the first one
cdr <- cdr_raw[, .(deathdate = min(deathdate, na.rm = TRUE)), by = "lopnr"]
pop <- left_merge(pop, cdr, id_vars = "lopnr")

dead_before_indexdate <- truethy(pop$deathdate < pop$index_date)

if (any(dead_before_indexdate)) {
    print(glue("deaths before index date should not be present (add CRD)
    Removing {sum(dead_before_indexdate)} rows where deathdate occurrs before proceduredate
    "))
    pop <- pop[!dead_before_indexdate]
}
```

2. #### Migrations

Added from FoB. Used for cencoring. Only emmigration after `index_date` are of interest

```{r add migration}

migrations_after_first <- migrations_raw[index_date < migrationdate, ]
setorder(migrations_after_first, lopnr, migrationdate)
migrations_after_first <- migrations_after_first[!duplicated(lopnr)]

# Only care for people that migrate after index_date
emmigrations_after_first <- migrations_after_first[
    migration_type == "emmigration",
    .(lopnr, index_date, migrationdate)
]

pop <- left_merge(pop, emmigrations_after_first, by = c("lopnr", "index_date"))

# n percent with emmigration after index_date
n_per(!is.na(pop$migrationdate))
```

3. #### Recurrent disease (relapse)

Recurrent disease is assumed to have arissen 6 months prior date given in SCRCR
(relapse=1 from CRC-base, date variable; first of `relapse_loc_date` or `relapse_mets_date` from CRC-Base)

_Also identified from the patient register it is a bit of an issue that we have two sources of relapse identification for the exposed. I suggest we only use patreg_

```{r recurrent disease}
first_relapse_date_scrcr <- scrcr_processed %>%
    group_by(lopnr) %>%
    summarise(
        relapse_scrcr_date = min(relapse_loc_date, relapse_mets_date, na.rm = TRUE)
    )

first_relapse_date_patreg <- relapse_raw[relapse_opr == 1 | relapse_ipr == 1, .(
    relapse_patreg = any(truethy(relapse_opr | relapse_ipr)),
    relapse_patreg_date = min(indate, na.rm = TRUE)
),
by = "lopnr"
]

# add index date
first_relapse_date_patreg <- left_join(first_relapse_date_patreg, pop[, .(lopnr, index_date)], by = "lopnr")
first_relapse_date_patreg <- first_relapse_date_patreg[!is.na(index_date)]


# combine relapses from the two sources
# TODO change from here if we want to remove scrcr as a source for relapse cencoring!
first_relapse <- full_join(first_relapse_date_patreg, first_relapse_date_scrcr, by = "lopnr")
first_relapse[, relapse_scrcr := ifelse(!is.na(relapse_scrcr_date), TRUE, FALSE)]
first_relapse[, relapse_date := pmin(relapse_patreg_date, relapse_scrcr_date, na.rm = TRUE)]

pop <- left_merge(pop, first_relapse[, .(lopnr, index_date, relapse_date)], by = c("lopnr", "index_date"))

# Make sure there are no relapses before procedure date
relapse_before_index <- truethy(pop$relapse_date < pop$index_date)

if (any(relapse_before_index)) {
    print(glue("Removing {sum(relapse_before_index)} rows with relapse before index date (proceduredate) when adding relapses from patreg and SCRCR.
    Should they be able to have a relapse before the index_date in SCRCR?"))
    # print the exluded rows
    print(first_relapse[relapse_date < index_date] %>% mutate(lopnr = NA))
    pop <- pop[!relapse_before_index]
}
```

### Baseline characteristics

#### SCRCR only
All variables are assessed as prevalent at baseline
 - Body mass index (BMI) (weight/height2  from CRC-base, kg/m2)
 - Neoadjuvant radiotherapy (neoadj_rt=1 from CRC-base)
 - Minimal invasive surgery (a2_lapa=1 or a2_robotass=1 from SCRCR) _to be exposure instead?_
 - Converted to open (a2_konv=1 from SCRCR)
 - Stoma formation at resection (a2_skydd=1 or a2_perm=1 from SCRCR)
 - Intraoperative blood-loss (a2_blodn from SCRCR) __perop_bleed = a2_blodn__
 - Pathological tumor stage (tstage_path=1-13 from CRC-base)
 - Location of tumour; right (a2_tumlage=1-5) left (a2_tumlage=6-8) rectum (a2_tumlok_beskrivning=2)
 - Colorectal cancer surgery type (a2_optyp2=01-09 [surgery_type has been derived] from SCRCR)
 - Year of surgery (`proceduredate`)

 

```{r baseline scrcr, include=TRUE}

# BMI - already in data, make categories
pop[, bmi_cat := case_when(
    bmi < 25 ~ "<25",
    bmi < 30 ~ "25-30",
    bmi >= 30 ~ ">=30",
    TRUE ~ ""
)]

# neoadjuvant radiotherapy
pop[, bs_neoadj_radio := truethy(neoadj_rt)]

# minimal invasive surgery
pop[, bs_minimal_invasive_surgery := case_when(
    truethy(a2_lapa) | truethy(a2_robotass) ~ TRUE,
    TRUE ~ FALSE
)]
pop[, .(n_minimal_invasive_surgery = sum(bs_minimal_invasive_surgery)), by = .(a2_lapa, a2_robotass)]

# Converted to open
pop[, bs_convert_to_open := truethy(a2_konv)]

# stoma formation at resection
pop[, bs_stoma_resection := case_when(
    truethy(a2_skydd) | truethy(a2_perm) ~ TRUE,
    TRUE ~ FALSE
)]
pop[, .(n_stoma = sum(bs_stoma_resection)), by = .(a2_skydd, a2_perm)]

# Intraop blood loss
pop[, bs_perop_bleed := perop_bleed]

# patological tumour stage
pop[, bs_path_t_stage := pt]
pop[, bs_path_t_stage_cat := pt_cat]

# Location of tumour
pop$tumour_location <- define_tumour_location(pop$location)
pop$tumour_side <- define_tumour_side(pop$col_location)
pop[, bs_location_of_tumour := case_when(
    tumour_location == "rectum" ~ "rectum",
    tumour_side == "right" ~ "right",
    tumour_side == "left" ~ "left",
    tumour_location == "colon" ~ "colon-unknown side",
    TRUE ~ ""
)]
tabyl(pop$bs_location_of_tumour)
# Colorectal cancer surgery type (surgery_type)
pop$bs_surgery_type <- pop$surgery_type

# Surgery year
pop[, index_year := year(index_date)]
pop$index_year %>% tabyl()
```


#### Whole population

- Abdominal surgery from 1997 to before the colorectal cancer surgery (op= se list, codes from IPR) before diagdate_scrcr _all that starts with J_
- SBO diagnosis (dia1-dia30=K56.5, K56.6, K56.7 from IPR) from 1997 to 3 months within CRC diagnosis (a1_diagnosdatum from SCRCR)  _added a washout period here of 0.5 months to avoid_
- Abdominal pain: First hospitalisation for abdominal pain (dia1-dia30=R10.0, R10.3, R10.4 from IPR) from 1997 to 3 months before CRC operation (diagdate_scrcr from CRC-Base).
- Education (from LISA)
- age categories (<65, 65-75 >75)
- Charlson (CCI)

```{r baseline all abdominal surgeries before procedure date}
# Adominal surgeries before procedure date
abd_surgeries_before <- previous_abd_surgeries_raw[indate < diagdate_scrcr & surgerycodes %like% "^J| J"]

# Present frequency of J codes. only 10 most frequent
all_codes <- abd_surgeries_before$surgerycodes %>%
    str_split(" ") %>%
    unlist()
j_codes <- all_codes[str_detect(all_codes, "^J")]
tabyl(j_codes) %>%
    arrange(-n) %>%
    head(10)

# Diagdate ibstead of indexdate to not confuse with exposure surgery
first_abd_surgeries_before <- abd_surgeries_before[, .(bs_previous_abd_surgery = 1), by = c("lopnr", "diagdate_scrcr")]
pop <- left_merge(pop, first_abd_surgeries_before, by = c("lopnr", "diagdate_scrcr"))
```

```{r baseline all sbo_3m_before}
# sbo 3 months before crcdiagnosis (ileus dataset)
# Already added index date to data
ileus_3m_before <- ileus_raw[indate < diagdate_scrcr & ileus == 1]
ileus_3m_before <- ileus_3m_before[
    ,
    months_before_diagdate := time_length(as_date(diagdate_scrcr) - as_date(indate), unit = "months")
][
    , .(lopnr, index_date, indate, diagdate_scrcr, ileus, months_before_diagdate)
]
ileus_3m_before <- ileus_3m_before[months_before_diagdate <= 3]
first_ileus_3m_before_diagdate <- ileus_3m_before[, .(
    date_ileus_within_3m = min(indate), bs_previous_sbo_3m_before_diagdate = 1
),
by = c("lopnr", "index_date")
]

# ileus distribution
qplot(first_ileus_3m_before_diagdate$date_ileus_within_3m) + theme_minimal()

pop <- left_merge(
    pop,
    first_ileus_3m_before_diagdate[, .(lopnr, index_date, bs_previous_sbo_3m_before_diagdate)],
    id_vars = c("lopnr", "index_date")
)
```

```{r abdominal pain within 3m before diagdate}
# COntains both hernia and pain
abpain_3m_before <- abd_hernia_raw[abpain == 1]

# Create months before diagdate and select variables
abpain_3m_before <- abpain_3m_before[
    ,
    months_before_diagdate := time_length(as_date(diagdate_scrcr) - as_date(indate), unit = "months")
][
    , .(lopnr, index_date, indate, diagdate_scrcr, abpain, months_before_diagdate)
]
# Within 3 months
abpain_3m_before <- abpain_3m_before[months_before_diagdate <= 3]

first_abpain_3m_before_diagdate <- abpain_3m_before[, .(
    date_abpain_within_3m = min(indate),
    bs_previous_abpain_3m_before_diagdate = 1
), by = c("lopnr", "index_date")]

# abpain distribution
if (nrow(first_abpain_3m_before_diagdate) == 0) {
    print("Message not able to identifty abpain from inpatient register")
} else {
    qplot(first_abpain_3m_before_diagdate$date_abpain_within_3m) + theme_minimal()
}

pop <- left_merge(
    pop,
    first_abpain_3m_before_diagdate[, .(lopnr, index_date, bs_previous_abpain_3m_before_diagdate)],
    id_vars = c("lopnr", "index_date")
)
```

```{r education}
most_recent_education <- get_education_from_lisa(lisa_raw, pop)
setDT(most_recent_education)

pop <- left_merge(pop, most_recent_education[, .(lopnr, index_date, education_cat)], by = c("lopnr", "index_date"))

# proportion missing education
n_per(pop$education_cat == "")
```


```{r cci}

cci <- rbindlist(list(
    cci_exposed_raw[, .(lopnr, diagdate_scrcr, CCI_weighted)],
    cci_comparators_raw[, .(lopnr, diagdate_scrcr, CCI_weighted)]
))
colnames(cci) <- str_to_lower(colnames(cci))
cci_max <- cci[, .(cci_weighted = max(cci_weighted)), by = c("lopnr", "diagdate_scrcr")]

pop <- left_merge(pop, cci_max, by = c("lopnr", "diagdate_scrcr"))
pop[is.na(cci_weighted), cci_weighted := 0]
```

```{r age cat}
pop[, age_cat := case_when(
    diagage < 65 ~ "<65",
    diagage >= 65 & diagage <= 75 ~ "65-75",
    diagage > 75 ~ ">75"
)]
```


#### Postoperation variables
Occurs after exposure.
Used in __Postoperation table__
All variables prefixed wih `postop_`

- Re-operation (a4_reop from SCRCR)
- Fascial dehiscence (a4_reoprup from SCRCR)
- Anastomotic leak (a4_anakomp from SCRCR)

```{r postoperation variables}
pop[, `:=`(
    postop_reoperation = truethy(reoperation), # derived from a4_reop
    postop_fascial = truethy(a4_reoprup),
    postop_al = truethy(a4_anakomp)
)]
```

- abdominal hernia

Incarcerated abdominal hernia (dia1-dia30= K40.0, K40.3, K41.0, K41.3, K42.0, K43.0, K43.3, K43.6, K44.0, K45.0, K46.0 as primary or secondary codes in the IPR)

```{r hernia}

hernia <- abd_hernia_raw[abhernia == 1 & index_date < indate]
hernia_first <- hernia[,
    .(postop_hernia = any(abhernia == 1), hernia_date = min(indate)),
    by = c("lopnr", "index_date")
]

# First hernia date after index date. But this is not the same time as sbo
hernia_first$hernia_date %>% qplot()

# Add to population
pop <- left_merge(pop, hernia_first[, .(lopnr, index_date, postop_hernia)], id_vars = c("lopnr", "index_date"))

# proportioni with hernia before index date
n_per(pop$postop_hernia)
```


### Outcomes

- `sbo_scrcr` u5_ileus_v_rde=1 (u0_besoksdatum for date sbo)
- `sbo_patreg` dia k56.5-7 (indatum for date sbo) found in uk_ileus dataset

```{r outcome sbo}

# sbo_scrcr (will be missing for all comparators)
sbo_scrcr <- scrcr_processed[u5_ileus == 1, .(
    sbo_scrcr = any(truethy(u5_ileus)),
    sbo_scrcr_date = min(u0_besoksdat)
), by = "lopnr"]


pop <- left_merge(pop, sbo_scrcr, id_vars = "lopnr")
if (any(pop$index_date >= pop$sbo_scrcr_date, na.rm = TRUE)) {
    warning("sbo_scrcr_date before index_date at sbo outcome. This should not happen as only the first scrcr should be included. Need to investigate")
}
```

```{r sbo patreg}
# sbo_patreg
# ileus = SBO

# Only keep ileus (also contains hernia and abpain icds)
# All codes are before diagdate
# only after index_date
ileus_after <- ileus_raw[ileus == 1 & index_date < indate]

# first one sbo from patreg
ileus_first <- ileus_after[, .(
    sbo_patreg = any(ileus == 1),
    sbo_patreg_date = min(indate, na.rm = TRUE)
), by = c("lopnr", "index_date")]

pop <- left_merge(pop, ileus_first, id_vars = c("lopnr", "index_date"))
```


```{r sbo summary}
pop[, .(sbo_scrcr = n_per(sbo_scrcr), sbo_patreg = n_per(sbo_patreg)), by = case]
```


#### 2. SBO-surgery 

__Assuming it does not require a prios SBO diagnosis__

- u5_ileus_reop_v_rde eq 1 (u0_besoksdatum for date sbo surgery)
- alt1 k56.5-7 plus op=JAP00, JAP01, JFK00, JFK01, JFK10, JFK20, JFK96, JFK97,
- alt2 k56.5-7 plus op= JAP00, JAP01, JFB00, JFB01, JFB10, JFB13, JFB20, JFB21, 
                        JFB96, JFB97, JFC00, JFC01, JFC10, JFC11, JFC20, JFC21,
                        JFF10, JFF11, JFF13, JFK00, JFK01, JFK10, JFK20, JFK96, JFK97 

```{r outcome sbo-surgery}

# SBO surgery scrcr
scrcr_surgery_scrcr <- scrcr_processed %>%
    filter(u5_ileus_reop == 1) %>%
    group_by(lopnr) %>%
    summarise(
        sbo_surgery_scrcr = any(u5_ileus_reop == 1, na.rm = TRUE),
        sbo_surgery_scrcr_date = min(u0_besoksdat, na.rm = TRUE)
    )

pop <- left_merge(pop, scrcr_surgery_scrcr, id_vars = "lopnr")
```


```{r sbo surgeries patreg}
# Must have ileus diagnosis and one of the codes. Must occurr after index date
ileus_only <- ileus_raw[ileus == 1 & index_date < indate]

# Find all rows with relevant sbo operational code
print(definitions$procedure_codes[1:2])
sbo_op_pattern1 <- definitions$procedure_codes[["sbo_operation_alt1"]] %>% str_c(collapse = "| ?")
sbo_op_pattern2 <- definitions$procedure_codes[["sbo_operation_alt2"]] %>% str_c(collapse = "| ?")

ileus_only[, `:=`(
    sbo_op1 = str_detect(surgerycodes, sbo_op_pattern1),
    sbo_op2 = str_detect(surgerycodes, sbo_op_pattern2)
)]

# definition2 is more loose so contains the most hits
ileus_only[, .(
    rows = .N,
    sbo_op1 = n_per(sbo_op1),
    sbo_op2 = n_per(sbo_op2),
    op1_and_op2 = n_per(sbo_op1 & sbo_op2),
    op1_or_op2 = n_per(sbo_op1 | sbo_op2)
)] %>% t()

sbo_surgeries <- ileus_only[sbo_op1 == 1 | sbo_op2 == 1]


# Find rows with ileus icd code
# looks through all the dia columns and returns a vector with one value per row indicating if any ileus icds are present
# this is already done in the ileus dataset but can keep for reality check
ileus_pattern <- definitions$icd_codes$ileus %>% str_c(collapse = "|")
sbo_surgeries$ileus_dia <- find_pattern_in_columns(select(sbo_surgeries, contains("dia")), ileus_pattern)

# proportion of rows with ileus dia
n_per(sbo_surgeries$ileus_dia)
# only keep the rows with ileus dia
sbo_surgeries <- sbo_surgeries[ileus_dia == TRUE]


# first occurrence for each code
first_sbo_surgeries <- sbo_surgeries %>%
    group_by(lopnr, index_date) %>%
    summarise(
        .groups = "drop",
        sbo_surgery_patreg1 = any(sbo_op1 == TRUE),
        sbo_surgery_patreg_date1 = min(indate[sbo_op1 == TRUE]),
        sbo_surgery_patreg2 = any(sbo_op2 == TRUE),
        sbo_surgery_patreg_date2 = min(indate[sbo_op2 == TRUE])
    )
nrow(first_sbo_surgeries)

pop <- left_merge(pop, first_sbo_surgeries, by = c("lopnr", "index_date"))
```

```{r sbo_surgery summary}
pop[, .(
    sbo_surgery_scrcr = n_per(sbo_surgery_scrcr),
    sbo_surgery_patreg1 = n_per(sbo_surgery_patreg1),
    sbo_surgery_patreg2 = n_per(sbo_surgery_patreg2)
), by = case]
```


### Secondary outcomes 
Is this right? Are these secondary outcomes? Should stoma closure be here too?

#### 1. Abdominal pain

First hospitalisation for abdominal pain (dia1-dia30=R10.0, R10.3, R10.4 from IPR) after CRC operation (proceduredate from CRC-Base). 
Cannot find any R1 codes in IPR.

```{r outcome abdominal pain}
abpain_after <- abd_hernia_raw[abpain == 1 & index_date < indate]

# Double check in IPR. even using just R1 does not works
# ipr_clean <- read_only_from_derived_csv("ipr_clean.csv")
# ipr$abpain <- find_pattern_in_columns(select(ipr, contains("dia")), "^R1")
# sum(ipr$abpain)
# abpain = ipr[abpain == TRUE]
# ipr %>% select(hdia:dia10, abpain)

abpain_after[, .(abpain = any(abpain), abpain_date = min(indate)), by = c("lopnr", "index_date")]
```


### End of follow up (eof)

SCRCR outcome variables will only be assessed for exposed (`sbo_scrcr_eof`). 

__Create__:

- `sbo_scrcr_eof`
- `sbo_scrcr_surgery_eof`
- `sbo_eof`
- `sbo_surgery_eof1`
- `sbo_surgery_eof2`

Which ever comes first of outcome_date, emmigration or death or Date of cancer recurrence(`relapse_date`) OR end of study (assumin `r params$end_of_study`)

```{r eof}
pop[, end_of_study := ymd(params$end_of_study)]
pop[, `:=`(
    sbo_scrcr_eof = pmin(sbo_scrcr_date, deathdate, migrationdate, relapse_date, end_of_study, na.rm = TRUE),
    sbo_surgery_scrcr_eof = pmin(deathdate, migrationdate, sbo_surgery_scrcr_date, relapse_date, end_of_study, na.rm = TRUE),
    sbo_eof = pmin(deathdate, migrationdate, sbo_patreg_date, relapse_date, end_of_study, na.rm = TRUE),
    sbo_surgery_eof1 = pmin(deathdate, migrationdate, sbo_surgery_patreg_date1, relapse_date, end_of_study, na.rm = TRUE),
    sbo_surgery_eof2 = pmin(deathdate, migrationdate, sbo_surgery_patreg_date2, relapse_date, end_of_study, na.rm = TRUE)
)]

pop %>%
    select(index_date, diagdate_scrcr, sbo_surgery_scrcr_date, deathdate, migrationdate, relapse_date, sbo_scrcr_eof, contains("eof")) %>%
    head()

# None 'date' date columns and their class
purrr::imap_dfr(select(pop, contains("date")), function(x, name) {
    if (!"Date" %in% class(x)) {
        tibble::tibble(
            name = name,
            class = str_c(class(x), collapse = ", ")
        )
    }
})
```

### Follow up time (fu) 
ie `time_since_proceduredate`
`index_date` to `end of follow up` in months.

```{r fu}

pop[, `:=`(
    sbo_scrcr_fu_m = time_length(as_date(sbo_scrcr_eof) - as_date(index_date), "months"),
    sbo_surgery_scrcr_fu_m = time_length(as_date(sbo_surgery_scrcr_eof) - as_date(index_date), "months"),
    sbo_fu_m = time_length(as_date(sbo_eof) - as_date(index_date), "months"),
    sbo_surgery_fu_m1 = time_length(as_date(sbo_surgery_eof1) - as_date(index_date), "months"),
    sbo_surgery_fu_m2 = time_length(as_date(sbo_surgery_eof2) - as_date(index_date), "months")
)]
```

```{r sbo fu summary}
plot_ly(data = pop) %>%
    add_trace(x = ~sbo_scrcr_fu_m, type = "histogram", name = "SBO scrcr") %>%
    add_trace(x = ~sbo_fu_m, type = "histogram", name = "SBO patreg")
```

```{r sbo surgery fu summary}
plot_ly(data = pop) %>%
    add_trace(x = ~sbo_surgery_scrcr_fu_m, type = "histogram", name = "SBO surgery scrcr") %>%
    add_trace(x = ~sbo_surgery_fu_m1, type = "histogram", name = "SBO surgery1") %>%
    add_trace(x = ~sbo_surgery_fu_m2, type = "histogram", name = "SBO surgery2")
```

### Follow up time categorized 

ie `time since proceduredate`
I dont get why start at 6 months here? _added <6m_

Three time periods (6-18 months, 18-41 months, 41-66 months)

__Done for__: 

- `sbo_fu_m` into `sbo_fu_cat_m`
- `sbo_surgery_fu_m` into `sbo_surgery_fu_cat_m`

```{r time_since_procedure}

# cat_followup <- function(x) {
#     cat_numeric(x, cutoffs = c(6, 18, 41, 66))
# }

# pop[, `:=`(
#     sbo_scrcr_fu_cat_m = cat_followup(sbo_scrcr_fu_m),
#     sbo_fu_cat_m = cat_followup(sbo_fu_m),
#     sbo_surgery_fu_cat_m1 = cat_followup(sbo_surgery_fu_m1),
#     sbo_surgery_fu_cat_m2 = cat_followup(sbo_surgery_fu_m2),
#     sbo_surgery_scrcr_fy_cat_m = cat_followup(sbo_surgery_scrcr_fu_m)
# )]
```

## Model variables
Additional variables used on SCRCR population only.
Outcome will be SBO for all
Data only available for SCRCR population

- exposure1 : method (a2_lapa)
- exposure2 : adjuvant_radiation (neoadj_rt)
- confounders
    - pT (1-3/4) `pt_group`
    - previous abdominal surgery `bs_prev_abd_surgeries`
    - tumour_level (a2_rscop [0-5, 6-10, 11-15])

```{r model variables}
# exposure1 method
pop[, method := case_when(
    a2_lapa == 1 ~ "lapa",
    TRUE ~ "open"
)]
tabyl(pop, method)

# exposure2 neoadj_rt
pop[, exp_neoadj_rt := truethy(neoadj_rt)]
tabyl(pop$exp_neoadj_rt)

# pT
pop[, pt_group := case_when(
    pt_cat %in% 2:4 ~ "1-3",
    pt_cat == 5 ~ "4",
    TRUE ~ "missing"
)]

# previous abdominal surgery (yes/no)
pop$bs_previous_abd_surgery <- truethy(pop$bs_previous_abd_surgery)
tabyl(pop$bs_previous_abd_surgery)

# Tumour level
pop[, tumour_level := case_when(
    a2_rscop %in% 0:5 ~ "0-5",
    a2_rscop %in% 6:10 ~ "6-10",
    a2_rscop %in% 11:15 ~ "11-15",
    TRUE ~ NA_character_
)]
tabyl(pop$tumour_level)

```


4. # Export analysisdata

```{r export}
print(glue("Writing analysisdata with : {nrow(pop)} rows and {ncol(pop)} columns to: project derived folder"))
write_to_project_derived(pop, "analysisdata.csv")
```
