==> 3_create_results.rmd <==
---
title: "Karlbom : Create results"
output:
  html_document:
    highlight: pygments
    theme: sandstone
    css: style.css
    toc: true
    code_folding: show
editor_options:
  chunk_output_type: inline
---

**Datum**: `r lubridate::today()`

```{r setup, include=FALSE, message=FALSE, warning=FALSE, echo=FALSE}
library(data.table)
library(dplyr)
library(stringr)
library(glue)
library(janitor)
library(here)
library(scales)
library(ggplot2)
library(lubridate)
library(rlang)
library(epitools) # CI for rate using exact method
library(cmprsk) # CIF and competing risk, crr(), timepoints(), cuminc()

source("crcbase_utils/helpers.R")
source("crcbase_utils/crud/setup_project_crud.R")
source("crcbase_utils/crud/read_from_variable_lists.R")
source("crcbase_utils/process/process_scrcr.R")
source("crcbase_utils/process/analysis.R")
test_for_k()

knitr::opts_chunk$set(
    # root.dir = '~/R/project',
    warning = FALSE,
    message = FALSE,
    warning = FALSE,
    echo = TRUE,
    include = TRUE
)

if (!exists("params")) {
    params <- list(
        project_name = project_name # from .Renviron
    )
}

# init tables to export
export_tables <- list()
```

# Read analysis data

Created in `2_crate_analysisdata()`. Look there for details.

```{r load analysis data}
ad_raw <- read_from_project_derived("analysisdata.csv")

# Remove unwanted columns
ad <- select_analysis_data(ad_raw)
# paste0(colnames(ad), collapse = " | ")

ad$tumour_location <- define_tumour_location(ad$location)
```

### Create factors
```{r add factor}
# exposed
ad[, case := factor(case, levels = c("exposed", "comparators"))]
tabyl(ad, case)
```

## Create variables

```{r}

# Age at proceduredate. Cannot find birthdate,
# using birthyear instead to estimate age at proceduredate
ad[, age_index_year := index_year - birthyear]

# Categorize follow up time
cat_followup <- function(x) {
    cat_numeric(x, cutoffs = c(6, 18, 41, 66))
}

ad[, `:=`(
    sbo_scrcr_fu_cat_m = cat_followup(sbo_scrcr_fu_m),
    sbo_fu_cat_m = cat_followup(sbo_fu_m),
    sbo_surgery_fu_cat_m1 = cat_followup(sbo_surgery_fu_m1),
    sbo_surgery_fu_cat_m2 = cat_followup(sbo_surgery_fu_m2),
    sbo_surgery_scrcr_fy_cat_m = cat_followup(sbo_surgery_scrcr_fu_m)
)]
```

# Results

## Table 1 : Baseline characteristics for the whole population

These should probably be seperated into two tables, one containing just the exposed presenting scrcr variables and a second one
also containing comparators presenting variables available for all.

```{r table 1}

table1_fun <- function(data, group = "") {
    tabl <- data[,
        .(
            n = .N,
            age_indexyear_mean = mean_sd(age_index_year),
            age_indexyear_median = median_iqr(age_index_year),
            women = n_per(female),
            cci_0 = n_per(cci_weighted == 0),
            cci_1 = n_per(cci_weighted == 1),
            cci_2 = n_per(cci_weighted == 2),
            cci_o2 = n_per(cci_weighted > 2),
            cci_weighted_mean = mean_sd(cci_weighted),
            cci_weighted_median = median_iqr(cci_weighted),
            education_u9 = n_per(education_cat == "<9"),
            education_9_to_12 = n_per(education_cat == "9-12"),
            education_o12 = n_per(education_cat == ">12"),
            previous_abdominal_surgery = n_per(bs_previous_abd_surgery),
            indexyear_2007_2009 = n_per(index_year %in% 2007:2009),
            indexyear_2010_2012 = n_per(index_year %in% 2010:2012),
            indexyear_2013_2016 = n_per(index_year %in% 2013:2016),
            other_baseline_char = "",
            previous_sbo_3m_bf_diagdate = n_per(bs_previous_sbo_3m_before_diagdate),
            previous_abpain_3m_bf_diagdate = n_per(bs_previous_abpain_3m_before_diagdate),
            indexyear_median = median_iqr(index_year),
            # after index date
            relaspe = n_per(!is.na(relapse_date)),
            death_during_followup = n_per(!is.na(deathdate))
        ),
        by = group
    ]

    tabl <- transpose_dt(tabl)

    return(tabl)
}

tab1_pop <- table1_fun(ad, group = "case")

table_dt(tab1_pop, caption = "table 1", title_row_names = FALSE)
```



## Table 2 : Baseline characteristics for SCRCR population

Event occurring between exposure and outcome. Only relevant for the scrcr population

```{r table 2 postop table}
table2_fun <- function(data, group = "") {
    tab_res <- data[, .(
        n = .N,
        bmi_under_25 = n_per(bmi_cat == "<25"),
        bmi_25_30 = n_per(bmi_cat == "25-30"),
        bmi_over_30 = n_per(bmi_cat == ">=30"),
        bmi_na = n_per(is_na(bmi_cat)),
        neoadj_radiotherapy = n_per(bs_neoadj_radio),
        method_lapa = n_per(method == "lapa"),
        procedure = "surgery_type",
        st_il_res = n_per(surgery_type == 1),
        st_right_hem = n_per(surgery_type == 2),
        st_trans_res = n_per(surgery_type == 3),
        st_left_hem = n_per(surgery_type == 4),
        st_sigmoid_res = n_per(surgery_type == 5),
        st_total = n_per(surgery_type == 6),
        st_lower_ar = n_per(surgery_type == 7),
        st_rectum_amp = n_per(surgery_type == 8),
        st_hart = n_per(surgery_type == 9),
        st_tem = n_per(surgery_type == 10),
        st_local_ex = n_per(surgery_type == 11),
        st_laparatomi = n_per(surgery_type == 12),
        st_other = n_per(surgery_type == 13),
        st_appendectomy = n_per(surgery_type == 14),
        stoma_resection = n_per(bs_stoma_resection),
        perop_bleed_u150 = n_per(bs_perop_bleed < 150),
        perop_bleed_150_300 = n_per(bs_perop_bleed >= 150 & bs_perop_bleed < 300),
        perop_bleed_o300 = n_per(bs_perop_bleed > 300),
        perop_bleed_na = n_per(is.na(bs_perop_bleed)),
        perop_bleed_median = median_iqr(bs_perop_bleed),
        perop_bleed_mean = mean_sd(bs_perop_bleed),
        stage_0 = n_per(stage == 0),
        stage_1 = n_per(stage == 1),
        stage_2 = n_per(stage == 2),
        stage_3 = n_per(stage == 3),
        stage_4 = n_per(stage == 4),
        stage_missing = n_per(is.na(stage)),
        pt_cat1_3 = n_per(bs_path_t_stage_cat %in% 2:4), # 0-3 and 4 is wanted. But contains 1-6
        pt_cat4 = n_per(bs_path_t_stage_cat == 5),
        pt_cat0 = n_per(bs_path_t_stage_cat == 0),
        pt_catx = n_per(bs_path_t_stage_cat == 6),
        adjuvant_chemo = n_per(adj_ct),
        minimal_invasive_surgery = n_per(bs_minimal_invasive_surgery),
        convert_to_open = n_per(bs_convert_to_open),
        tumour_location_left_colon = n_per(bs_location_of_tumour == "left"),
        tumour_location_right_colon = n_per(bs_location_of_tumour == "right"),
        tumour_location_rektum = n_per(bs_location_of_tumour == "rektum"),
        tumour_location_undefined = n_per(bs_location_of_tumour == ""), # 9 = undefined
        anastomotic_site_ilecolic = n_per(bs_surgery_type %in% 1:2),
        anastomotic_site_colocolic = n_per(bs_surgery_type %in% 3:5),
        anastomotic_site_coloilorec = n_per(bs_surgery_type %in% 6:7)
    ), by = group]

    tab_res <- transpose_dt(tab_res)

    return(tab_res)
}


table2_fun(ad[case == "exposed"], group = "tumour_location") %>%
    table_dt(caption = "Baseline characteristics for the SCRCR population")
```

## Table 3: Postoperation variables

For scrcr population only

```{r postoperationtable}
ad[case == "exposed", .(
    n = .N,
    reoperation = n_per(postop_reoperation),
    fascial_del = n_per(postop_fascial),
    al = n_per(postop_al),
    abdominal_hernia = n_per(truethy(postop_hernia))
), by = "tumour_location"] %>%
    transpose_dt() %>%
    table_dt(caption = "Postoperation variables - SCRCR population only")
```

## Validation

For exposed only.

- Estimate sensitivity (true positive/true positive + false negative) of SBO in SCRCR compared to inpatient register.
- Estimate sensitivty of SBO surgery in SCRCR compared to two inpatient algorithms.
- The inpatient register is the gold standard.
- PPV as well (true positive/true positive + false positive)

### Validation of SBO (table2)

For exposed only. Validation of sbo and sbo surgery in scrcr with IPR as the gold standard.

```{r validation sbo}

# Sensitivity,
sensitivity <- function(test, gold) {
    true_positives <- sum(truethy(test == TRUE & gold == TRUE))
    all_positives <- sum(truethy(gold == TRUE))
    estimate <- scales::percent(true_positives / all_positives)
    glue("{true_positives}/{all_positives} ({estimate})")
}

ppv <- function(test, gold) {
    true_positives <- sum(truethy(test == TRUE & gold == TRUE))
    false_positives <- sum(truethy(test == TRUE & gold == FALSE))
    estimate <- scales::percent(true_positives / (true_positives + false_positives))
    glue("{true_positives}/{false_positives} ({estimate})")
}

ad[case == "exposed", .(
    n = .N,
    n_sbo_scrcr = n_per(sbo_scrcr),
    n_sbo_patreg = n_per(sbo_patreg),
    sensitivty = sensitivity(sbo_scrcr, sbo_patreg),
    ppv = ppv(sbo_scrcr, sbo_patreg)
), by = "sbo_fu_cat_m"] %>%
    arrange(sbo_fu_cat_m) %>%
    table_dt(caption = "Sensitivity and PPV for SBO in SCRCR compared to patreg")
```

_Conclusion_: Sensitivity of SBO in scrcr is low (when patreg is the gold standard)  

### Validation of SBO surgery (Table 2)

_correspondence?_
For exposed only.

```{r validation sbo surgery}

# Definition 1
ad[case == "exposed", .(
    n = .N,
    n_sbo_sugery_scrcr = n_per(sbo_surgery_scrcr),
    n_sbo_surgery_patreg1 = n_per(sbo_surgery_patreg1),
    sensitivty1 = sensitivity(sbo_surgery_scrcr, sbo_surgery_patreg1),
    ppv1 = ppv(sbo_surgery_scrcr, sbo_surgery_patreg1)
), by = "sbo_surgery_fu_cat_m1"] %>%
    arrange(sbo_surgery_fu_cat_m1) %>%
    table_dt(caption = "Sensitivity and PPV for SBO surgery in SCRCR compared to inpatient algorithm1")

# Definition 2
ad[case == "exposed",
    .(
        n = .N,
        n_sbo_surgery_scrcr = n_per(sbo_surgery_scrcr),
        n_sbo_surgery_patreg2 = n_per(sbo_surgery_patreg2),
        sensitivty2 = sensitivity(sbo_surgery_scrcr, sbo_surgery_patreg2),
        ppv2 = ppv(sbo_surgery_scrcr, sbo_surgery_patreg2)
    ),
    by = "sbo_surgery_fu_cat_m2"
] %>%
    arrange(sbo_surgery_fu_cat_m2) %>%
    table_dt(caption = "Sensitivity and PPV for SBO surgery in SCRCR compared to inpatient algorithm2")
```

_Conclusion_: 

- PPV is 100% for algorithm2 (no false negatives in the SCRCR, they are all found in patreg too). SCRCR does not really add anything.
- Have these definition from patreg been validated elsewhere?

## Part 1: colorectal cancer as the exposure

### Proportional hazard assumption

The proportional hazard assumption was tested taking the ineraction term of log of follow up and exposure

```{r proportional hazard assumption}
prop_fit <- survival::coxph(Surv(sbo_fu_m, sbo_patreg) ~ case + case:log(sbo_fu_m), data = ad)
prop_fit %>% broom::tidy()

# Confirm with cox.zph (if global or the exposure has a p <0.05, the ph assumption is not valid)
global_fit <- survival::coxph(Surv(sbo_fu_m, sbo_patreg) ~ case, data = ad)
survival::cox.zph(global_fit)
```

```{r ph assumption sbo_patreg}
prop_fit <- survival::coxph(Surv(sbo_surgery_fu_m2, sbo_surgery_patreg2) ~ case + case:log(sbo_surgery_fu_m2), data = ad)
prop_fit %>% broom::tidy()
```
- The interaction term between the exposure variable and the log of follow up are below <0.05
- PH assumption does not hold for SBO nor SBO surgery



### Kapplan Meier

```{r kapplan meier}
kapplan_meier(ad, sbo_patreg, sbo_fu_m, exp_var = case)
kapplan_meier(ad, sbo_surgery_patreg2, sbo_surgery_fu_m2, exp_var = case)
```

### Incidence and HR of SBO

- Crude incidence of SBO for the two groups. 
- unadjusted HRs

```{r incidence of sbo}
confounders_base <- c("diagage", "female", "index_year")
bind_rows(
    inc_hr_row(ad, sbo_patreg, sbo_fu_m,
        exposure = case,
        ref = "comparators" # , confounders = confounders_base
    ),
    inc_hr_row(ad, sbo_surgery_patreg2, sbo_surgery_fu_m2,
        exposure = case, ref = "comparators" # , confounders = confounders_base
    )
) %>%
    table_dt(caption = "Crude incidence rate and hazard ratio of SBO and SBO surgery in colorectal cancer patientes (exposed) and general population comparators (unexposed)")
```

- Incidence per 1000 person years.
- Incidence CI estimated using the exact method (ref).


### Cumulative Incidence function (CIF)

- Competing risk = 'death'
- Cumulative incidence is estimated at 18, 44, 66 months

```{r cif status}
# Competing risk of death taken into account.
# Fine and Gray (cmprsk)
# code status according to 1 = sbo, 2 = alive/cencored, 3 = dead (competing event)


ad[, status_sbo := case_when(
    sbo_eof == sbo_patreg_date ~ 1,
    sbo_eof == deathdate ~ 2, # competing event
    TRUE ~ 0 # cencored (emmigration, end of study, relapse)
)]


ad[, status_sbo_surgery := case_when(
    sbo_surgery_eof2 == sbo_surgery_patreg_date2 ~ 1,
    sbo_surgery_eof2 == deathdate ~ 2,
    TRUE ~ 0 # cencored (emmigration, end of study, relapse)
)]
```


```{r cif sbo}
cif_sbo <- cmprsk::cuminc(
    ftime = ad$sbo_fu_m,
    fstatus = ad$status_sbo,
    group = ad$case,
    cencode = 0
)

# Create the cif table
cif_table(cif_obj = cif_sbo, timepoints = c(18, 41, 66)) %>%
    table_dt(caption = "Cumulative incidence of SBO in colorectal cancer patients (exposed) and general population comparators (unexposed)
 under the competing risk of death")
```

```{r}
plot(cif_sbo)
cif_sbo[1:2]
```
```{r plot cif}

survminer::ggcompetingrisks(cif_sbo, multiple_panels = F) +
    ggtitle("") +
    scale_linetype_manual(
        name = "Cohort:",
        values = c("dashed", "solid"),
        labels = c("Comparators", "Exposes")
    ) +
    scale_color_manual(
        name = "Event:",
        values = c("black", "darkgray"),
        labels = c("SBO", "Death")
    ) +
    scale_y_continuous(name = "Cumulaitve incidence (%)") +
    scale_x_continuous(name = "Follow up time (months)", breaks = c(0, 18, 41, 66, 125)) +
    theme(
        legend.position = c(0.2, 0.8),
        legend.title = element_text(face = "bold")
    )
```
__Main event (sbo) only__

```{r plot cif2}

survminer::ggcompetingrisks(cif_sbo[1:2], multiple_panels = F) +
    ggtitle("") +
    scale_linetype_manual(
        name = "Cohort:",
        values = c("dashed", "solid"),
        labels = c("Comparators", "Exposes")
    ) + 
    scale_color_manual(
        name = "Event:",
        values = c("black"),
        labels = c("SBO")
    ) +
    scale_y_continuous(name = "Cumulaitve incidence (%)") +
    scale_x_continuous(name = "Follow up time (months)", breaks = c(0, 18, 41, 66, 125)) +
    theme(
        legend.position = c(0.2, 0.8),
        legend.title = element_text(face = "bold")
    )
```

## Part 2: SCRCR population only

Only for SCRCR population. 
Two different exposures identified in SCRCR and the rest of the SCRCR population are included in the unexposed cohort.

### Model 1 : Lapa

- Study population: All colorectal cancer patientes in the study
- Exposure : Lapa
- Confounders: pT, BMI, earlier abdominal surgery, time period, sex, _age_

```{r index_period}
ad[, index_period := case_when(
    index_year %in% 2007:2009 ~ "2007:2009",
    index_year %in% 2010:2012 ~ "2010:2012",
    index_year %in% 2013:2016 ~ "2013:2016",
)]
```


```{r  model Lapa}
confounders1 <- c("diagage", "female", "pt_group", "bmi_cat", "bs_previous_abd_surgery", "index_period")

bind_rows(
    inc_hr_row(
        x = ad[case == "exposed"],
        outcome_in = sbo_patreg,
        persontime_in = sbo_fu_m,
        exposure_var = method, 
        ref = "open",
        confounder = confounders1,
        crude_hr = TRUE,
        print_summary = TRUE
    ),
    inc_hr_row(
        x = ad[case == "exposed"],
        outcome_in = sbo_surgery_patreg2,
        persontime_in = sbo_surgery_fu_m2,
        exposure_var = method, 
        ref = "open",
        confounder = confounders1,
        crude_hr = TRUE,
        print_summary = TRUE
    )
) %>%
    table_dt(caption = glue("Model 1: Lapa vs open, confounders={paste(confounders1, collapse = ', ')}"))
```

```{r model1 kapplan meier}
kapplan_meier(ad[case == "exposed"], sbo_patreg, sbo_fu_m, method)
kapplan_meier(ad[case == "exposed"], sbo_surgery_patreg2, sbo_surgery_fu_m2, method)
```

### Model 2 : Neoadjuvant radiotherapy
- Study population: All rectal cancer patients in the study
- Exposure: Neoadjuvant radio therapy
- Confounders: pT (pT1-3 / pT4), time period, sew, age, tumour level

```{r model2: neoadj_radiotherapy}

# Tumour location, needed for stratification
ad$tumour_location <- define_tumour_location(ad$location)

confounders2 <- c("diagage", "female", "pt_group", "index_period", "tumour_level")

bind_rows(
    inc_hr_row(
        x = ad[case == "exposed" & tumour_location == "rectum"],
        outcome_in = sbo_patreg,
        persontime_in = sbo_fu_m,
        exposure_var = exp_neoadj_rt,
        confounder = confounders2,
        crude_hr = TRUE,
        print_summary = TRUE
    ),
    inc_hr_row(
        x = ad[case == "exposed" & tumour_location == "rectum"],
        outcome_in = sbo_surgery_patreg2,
        persontime_in = sbo_surgery_fu_m2,
        exposure_var = exp_neoadj_rt,
        confounder = confounders2,
        crude_hr = TRUE,
        print_summary = TRUE
    )
) %>%
    table_dt(caption = glue("Model 2: Neoadj radiation therapy in rectal cancer patients, confounders={paste(confounders2, collapse = ', ')}"))
```

```{r km model 2}
kapplan_meier(ad[case == "exposed" & tumour_location == "rectum"], sbo_patreg, sbo_fu_m, exp_neoadj_rt)
kapplan_meier(ad[case == "exposed" & tumour_location == "rectum"], sbo_surgery_patreg2, sbo_surgery_fu_m2, exp_neoadj_rt)
```
